implemented steady + unsteady CCND apps for DFG95 benchmarks
This commit replaces the old prototypic 'navier_stokes_ccnd_q2p1' app by two
variants of the CCND app, one for steady-state and another one for unsteady
simulations. Those variants are implemented in two separate source files,
namely 'ccnd_steady.cpp' and 'ccnd_unsteady.cpp', and some common functionality
has been outsourced to the 'ccnd_common.hpp' header, which is included by both
applications. Although the two source files contain a lot of duplicate code, it
was decided to separate the steady and unsteady code paths to keep the code
lean and focused on the task.

A quick summary of the main features of both variants:
* Isoparametric Q2/P1dc finite element discretisation in 2D and 3D with
  support for recursive paritioning.
* Newton-Multigrid-AmaVanka solver with adaptive residual control
  for the multigrid preconditioner to minimize solver times.
* Computation of various post-processing quantities required by the DFG95
  benchmarks, i.e. drag/lift coefficients and pressure differences, and
  some additional quantities like fluxes or velocity field norms.
* Support for quadruple precision, if the quadmath build is supplied.
  In this case, two additional build targets 'dfg95-ccnd-steady-qp' and
  'dfg95-unsteady-qp' are available.
* Logging of various timing and performance statistics.
* Needless to say, full support for MPI parallelization.

In addition to that, the unsteady variant offers the following features:
* Second-order BDF(2) time-stepping scheme with quadratic extrapolation
  of the previous time-steps to obtain an initial guess for the next one.
* Checkpoint-Restart system that allows to continue a previously aborted
  simulation.
* Automatic detection of steady-state by analysis of the velocity field
  time-derivative.

Both applications have been configured to solve the infamous DFG95
flow-around-a-cylinder benchmarks 1,2,3 in 2D and 3D.

Additionally, this commit also includes two Python scripts, 'bench2plots.py'
and 'bench3plots.py', which can be used to analyse the output logs generated by
the unstead ccnd variant as a post-processing step. These scripts will create
Matplotlib PyPlot scripts, which visualise various benchmark quantities, such
as e.g. drag and lift, over the whole simulation time or the periodic cycle.
Furthermore, these scripts also perform some basic analysis of those quantities
including an EOC analysis.

+++ kernel/geometry/export_vtk.hpp
@@ -282,7 +282,7 @@ namespace FEAT
             d[Index(3)*i+j] = double(0);
 
           for(int j(0); j < v.BlockSize; ++j)
-            d[Index(3)*i+Index(j)] = v(i)[j];
+            d[Index(3)*i+Index(j)] = double(v(i)[j]);
 
         }
         _vertex_vectors.push_back(std::make_pair(name, std::move(d)));
@@ -394,7 +394,7 @@ namespace FEAT
             d[Index(3)*i+j] = double(0);
 
           for(int j(0); j < v.BlockSize; ++j)
-            d[Index(3)*i+Index(j)] = v(i)[j];
+            d[Index(3)*i+Index(j)] = double(v(i)[j]);
 
         }
         _cell_vectors.push_back(std::make_pair(name, std::move(d)));
